requirements.txt
# Main packages
Flask~=2.0.2
flask-restx
Flask-CORS
Flask-SQLAlchemy
marshmallow~=3.14.1
PyJWT

# For test
pytest~=7.0.0
pytest-cov
SQLAlchemy~=1.4.31